Modif backEnd 3

+++ src/sample/Controller/Calcul.java
@@ -2,6 +2,8 @@ package sample.Controller;
 
 import sample.Model.BTS;
 
+import java.util.*;
+
 public final class Calcul {
 
     public Calcul() {
@@ -9,23 +11,23 @@ public final class Calcul {
     }
 
     //calcul distance
-    public static double distance(double x1, double y1, double x2, double y2){
-        return Math.sqrt(
-                Math.pow(Math.abs(x2-x1),2)+
-                        Math.pow(Math.abs(y2-y1),2));
+    public static float distance(double x1, double y1, double x2, double y2) {
+        return (float) Math.sqrt(
+                Math.pow(Math.abs(x2 - x1), 2) +
+                        Math.pow(Math.abs(y2 - y1), 2));
     }
 
     //calcul Friiz
-    public static double calculReceivingPower(BTS bts, double gainReceiving, double lamda, double distance){
+    public static double calculReceivingPower(BTS bts, double gainReceiving, double lamda, float distance) {
 
-        double fspl = 20 * Math.log(lamda / (4 * Math.PI *distance));
+        double fspl = 20 * Math.log(lamda / (4 * Math.PI * distance));
 
         return bts.getPowerTransmitting() + gainReceiving + bts.getGainTransmitting() + fspl;
     }
 
     //wave length calcul
-    public static double lamda(Long frequency){
+    public static double lamda(Long frequency) {
 
-        return (Math.pow(10,8) * 3) / frequency;
+        return (Math.pow(10, 8) * 3) / frequency;
     }
 }
